chore(src): reorganize Menu elements
